Rename gurobi specific files and add GLPK flag
