+++ src/optimize/constraint/dynamics/gurobi_final_charge.py
@@ -5,7 +5,7 @@ from constraint import Constraint
 
 ##===============================================================================
 #
-class FinalCharge(Constraint):
+class GBFinalCharge(Constraint):
     ##=======================================================================
     # PUBLIC
