+++ src/optimize/constraint/dynamics/gurobi_initial_charge.py
@@ -5,7 +5,7 @@ from constraint import Constraint
 
 ##===============================================================================
 #
-class SpaceBigO(Constraint):
+class GBInitialCharge(Constraint):
     ##=======================================================================
     # PUBLIC
 
@@ -20,16 +20,16 @@ class SpaceBigO(Constraint):
     #           NONE
     #
     def constraint(self, model, params, d_var, i, j):
-        model.update()
         # Extract parameters
-        Q = self.params['Q']
-        s = self.params['s']
+        G     = self.params['Gamma']
+        alpha = self.params['alpha']
+        kappa = self.params['kappa']
 
         # Extract decision vars
-        delta = self.d_var['delta']
-        v     = self.d_var['v']
+        eta = self.d_var['eta']
 
-        if i != j:
-            model.addConstr(v[i] - v[j] - s[i] - (delta[i][j] - 1)*Q >= 0, \
-                                      name="{0}_{1}_{2}".format(self.name,i,j))
+        if alpha[i] > 0:
+            model.addConstr(alpha[i]*kappa[G[i]] == eta[i], \
+                                    name="{0}_{1}".format(self.name, i))
+
         return
