+++ src/optimize/constraint/dynamics/gurobi_min_charge_propagation.py
@@ -5,7 +5,7 @@ from constraint import Constraint
 
 ##===============================================================================
 #
-class ChargePropagation(Constraint):
+class GBMinChargePropagation(Constraint):
 	##=======================================================================
 	# PUBLIC
 
@@ -21,17 +21,19 @@ class ChargePropagation(Constraint):
 	#
 	def constraint(self, model, params, d_var, i, j):
 		# Extract parameters
-		Q   = self.params['Q']
-		gam = self.params['gamma']
-		r   = self.params['r']
-		l   = self.params['l']
+		G     = self.params['Gamma']
+		Q     = self.params['Q']
+		gam   = self.params['gamma']
+		l     = self.params['l']
+		kappa = self.params['kappa']
+		nu    = self.params['nu']
+		r     = self.params['r']
 
 		# Extract decision vars
 		eta = self.d_var['eta']
 		g   = self.d_var['g']
 
-		if gam[i] >= 0:
-			model.addConstr(eta[i] + sum(g[i][q]*r[q] for q in range(Q)) - l[i] == eta[gam[i]], \
-										  name="{0}_{1}".format(self.name,i))
+		model.addConstr(eta[i] + sum(g[i][q]*r[q] for q in range(Q)) - l[i] >= nu*kappa[G[i]], \
+										name="{0}_{1}".format(self.name,i))
 
 		return
