+++ src/optimize/constraint/packing/gurobi_sigma.py
@@ -5,27 +5,25 @@ from constraint import Constraint
 
 ##===============================================================================
 #
-class ValidDepartureTime(Constraint):
+class GBSigma(Constraint):
     ##=======================================================================
     # PUBLIC
 
     ##-----------------------------------------------------------------------
     # Input:
-    #           m     : Gurobi model
-    #           params: Model parameters
-    #           d_var : Model decision variables
-    #           i     : constraint id
+    #       m     : Gurobi model
+    #       params: Model parameters
+    #       d_var : Model decision variables
+    #       i     : constraint id
     #
     # Output:
     #           NONE
     #
     def constraint(self, model, params, d_var, i, j):
-        # Extract parameters
-        t = self.params['t']
-
         # Extract decision vars
-        c = self.d_var['c']
+        sigma = self.d_var['sigma']
 
-        model.addConstr(c[i] <= t[i], name="{0}_{1}".format(self.name,i))
-
+        if i != j:
+            model.addConstr(sigma[i][j] + sigma[j][i] <= 1, \
+                            name="{0}_{1}_{2}".format(self.name,i,j))
         return
