+++ src/optimize/constraint/packing/gurobi_valid_end_time.py
@@ -5,7 +5,7 @@ from constraint import Constraint
 
 ##===============================================================================
 #
-class ValidEndTime(Constraint):
+class GBValidEndTime(Constraint):
     ##=======================================================================
     # PUBLIC
