+++ src/optimize/constraint/packing/gurobi_valid_initial_time.py
@@ -5,7 +5,7 @@ from constraint import Constraint
 
 ##===============================================================================
 #
-class MinChargePropagation(Constraint):
+class GBValidInitialTime(Constraint):
 	##=======================================================================
 	# PUBLIC
 
@@ -21,19 +21,10 @@ class MinChargePropagation(Constraint):
 	#
 	def constraint(self, model, params, d_var, i, j):
 		# Extract parameters
-		G     = self.params['Gamma']
-		Q     = self.params['Q']
-		gam   = self.params['gamma']
-		l     = self.params['l']
-		kappa = self.params['kappa']
-		nu    = self.params['nu']
-		r     = self.params['r']
+		a = self.params['a']
 
 		# Extract decision vars
-		eta = self.d_var['eta']
-		g   = self.d_var['g']
+		u = self.d_var['u']
 
-		model.addConstr(eta[i] + sum(g[i][q]*r[q] for q in range(Q)) - l[i] >= nu*kappa[G[i]], \
-										name="{0}_{1}".format(self.name,i))
-
+		model.addConstr(a[i] <= u[i] , name="{0}_{1}".format(self.name,i))
 		return
