+++ src/optimize/optimizer.py
@@ -23,13 +23,15 @@ class Optimizer:
     ##===========================================================================
 
     ##---------------------------------------------------------------------------
-    # Input:
-    #       NONE
-    #
-    # Output:
-    #       NONE
     #
     def __init__(self):
+    """
+     Input:
+           NONE
+
+     Output:
+           NONE
+    """
         # Parse 'config/general.yaml'
         with open(r'config/general.yaml') as f:
                 file           = yaml.load(f, Loader=yaml.FullLoader)
@@ -37,6 +39,7 @@ class Optimizer:
                 self.verbose   = file['verbose']
                 self.lff       = file['load_from_file']
                 self.time_lim  = file['time_limit']
+                self.solver    = file['solver']
 
         # Initialize member variables
         self.dm         = DataManager()
